Add files for test

+++ ConversorTemp.java
@@ -1,14 +1,14 @@
 import java.util.*;
 
 public class ConversorTemp {
-    public float fahrenheitCelsius(float f) {
-        float c;
+    public double fahrenheitCelsius(double f) {
+        double c;
         c = (f-32)*5/9;
         return c;
     }
 
-    public float celsiusFahrenheit(float c) {
-        float f;
+    public double celsiusFahrenheit(double c) {
+        double f;
         f = (c*9/5)+32;
         return f;
     }
@@ -17,7 +17,7 @@ public class ConversorTemp {
         Scanner entrada = new Scanner(System.in);
         ConversorTemp calc = new ConversorTemp();
         int controle = -1;
-		float c, f; 
+		double c, f; 
 
 		System.out.print("CONVERSOR DE TEMPERATURA\n\n"); 
 		
@@ -30,13 +30,13 @@ public class ConversorTemp {
             switch (controle) {
                 case 1: 
                     System.out.print("Digite a temperatura (C): ");
-                    c = entrada.nextFloat();
+                    c = entrada.nextDouble();
                     System.out.print("\nTemperatura em Fahrenheint: ");
                     System.out.print(calc.celsiusFahrenheit(c) +" F\n\n");
                     break;
                 case 2:
                     System.out.print("Digite a temperatura (F): ");
-                    f = entrada.nextFloat();
+                    f = entrada.nextDouble();
                     System.out.print("\nTemperatura em Celsius: ");
                     System.out.print(calc.fahrenheitCelsius(f) + " C\n\n");
                     break;
